Make `isassigned` work with tuples
This commit adds a method `isassigned(v::Tuple, i::Integer)` which makes
that function work on `Tuple`s, the same way it does on `Vector`s.
(Since a `Tuple` cannot be created without being assigned, all that
is left is to check that the index is within bounds.)

This makes it easier to write generic code that works on both
`Vector`s and `Tuple`s.